# public/openapi.yaml
openapi: 3.1.1
info:
  title: YESorNOaaS API
  version: 2.2.6
  description: |
    Call `/api/yes` for `Yes!`, `/api/no` for `No!`, or `/api/random` for a
    random `Yes!` or `No!`.
    Static assets, `GET /version`, `GET /health`, and `GET /metrics` are served
    outside the answer routes. Legacy root share URLs with a `request` query
    parameter redirect to `/yes`, `/no`, or `/random`. Other unmatched paths
    return the HTML 404 page. Rate-limited requests return `429 text/plain`
    with a route hint.
paths:
  /:
    get:
      summary: Redirect legacy root share links to the matching answer page
      parameters:
        - name: request
          in: query
          required: true
          schema:
            type: string
        - name: answer
          in: query
          required: false
          schema:
            type: string
            enum: [yes, no, random]
      responses:
        '308':
          description: Redirect to the matching answer page
          headers:
            Location:
              schema:
                type: string
        '404':
          description: Root path without a request query parameter
          content:
            text/html:
              schema:
                type: string
  /version:
    get:
      summary: Return the package version as plain text
      responses:
        '200':
          description: Package version
          content:
            text/plain:
              schema:
                type: string
  /health:
    get:
      summary: Return service health and package version
      responses:
        '200':
          description: YESorNOaaS health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is draining connections during shutdown
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /metrics:
    get:
      summary: Return Prometheus metrics for scraping
      description: |
        Exposes HTTP service metrics and standard Node.js runtime metrics in
        Prometheus text format. Intended for Prometheus scraping. Exempt from
        rate limiting.
      responses:
        '200':
          description: Prometheus metrics exposition format
          content:
            text/plain:
              schema:
                type: string
  /api/yes:
    get:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    put:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    post:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    delete:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    options:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    head:
      responses:
        '200':
          description: No response body
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    patch:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    trace:
      responses:
        '200':
          $ref: '#/components/responses/YesResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
  /api/no:
    get:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    put:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    post:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    delete:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    options:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    head:
      responses:
        '200':
          description: No response body
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    patch:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    trace:
      responses:
        '200':
          $ref: '#/components/responses/NoResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
  /api/random:
    get:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    put:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    post:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    delete:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    options:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    head:
      responses:
        '200':
          description: No response body
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    patch:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
    trace:
      responses:
        '200':
          $ref: '#/components/responses/RandomResponse'
        '429':
          $ref: '#/components/responses/ThrottledResponse'
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
          const: YESorNOaaS
        version:
          type: string
  responses:
    YesResponse:
      description: YESorNOaaS yes answer
      content:
        text/plain:
          schema:
            type: string
            const: Yes!
    NoResponse:
      description: YESorNOaaS no answer
      content:
        text/plain:
          schema:
            type: string
            const: No!
    RandomResponse:
      description: YESorNOaaS random yes or no answer
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        text/plain:
          schema:
            oneOf:
              - type: string
                const: Yes!
              - type: string
                const: No!
    ThrottledResponse:
      description: Rate limit exceeded
      content:
        text/plain:
          schema:
            type: string
            const: 'Use /api/yes, /api/no, or /api/random. Example: GET /api/yes -> Yes!'
x-yesornoaas-unknown-routes:
  description: Unmatched request paths, except legacy root share redirects, return the HTML 404 page until throttled, then `429 text/plain` with a route hint.